+++ apps/prairielearn/src/tests/helperDb.ts
@@ -21,7 +21,7 @@ const POSTGRES_INIT_CONNECTION_STRING = 'postgres://postgres@localhost/postgres'
 const POSTGRES_DATABASE = 'pltest';
 const POSTGRES_DATABASE_TEMPLATE = 'pltest_template';
 
-export const postgresTestUtils = sqldb.makePostgresTestUtils({
+const postgresTestUtils = sqldb.makePostgresTestUtils({
   user: POSTGRES_USER,
   host: POSTGRES_HOST,
   defaultDatabase: 'postgres',
@@ -103,7 +103,7 @@ async function closeSql(): Promise<void> {
   await sqldb.closeAsync();
 }
 
-async function databaseExists(dbName: string): Promise<boolean> {
+export async function databaseExists(dbName: string): Promise<boolean> {
   const client = new pg.Client(POSTGRES_INIT_CONNECTION_STRING);
   await client.connect();
   const result = await client.query(
@@ -115,6 +115,7 @@ async function databaseExists(dbName: string): Promise<boolean> {
 }
 
 export async function setupDatabases({ configurePool = true }: { configurePool?: boolean } = {}) {
+  // In e2e tests, the template database should always exist.
   const templateExists = await databaseExists(POSTGRES_DATABASE_TEMPLATE);
   const dbName = getDatabaseNameForCurrentWorker();
   if (!templateExists) {
@@ -218,6 +219,7 @@ export async function dropTemplate(): Promise<void> {
 
 /**
  * Helper function for testing migrations.
+ *
  * @param params
  * @param params.name The name of the migration to test.
  * @param params.beforeMigration A function to run before the migration.